Deleted my drivers, moved my factories to use_case, deleted SearchFrame and removed usage in ChangePasswordView
