+++ src/main/java/use_case/login/LoginUseCaseFactory.java
@@ -1,17 +1,12 @@
-package app;
+package use_case.login;
 
 import interface_adapter.ViewManagerModel;
 import interface_adapter.change_password.LoggedInViewModel;
 import interface_adapter.login.LoginController;
 import interface_adapter.login.LoginPresenter;
 import interface_adapter.login.LoginViewModel;
-import use_case.login.LoginInputBoundary;
-import use_case.login.LoginInteractor;
-import use_case.login.LoginOutputBoundary;
-import use_case.login.LoginUserDataAccessInterface;
 import view.LoginView;
 
-// TODO: DELETE
 public final class LoginUseCaseFactory {
